perf: fix style
